CPC conversion to celery

+++ updater/create_databases/upload_new.py
@@ -31,31 +31,34 @@ def consolidate_cpc_classes(connection_string):
     for table_name in ['mainclass', 'subclass']:
         engine = create_engine(connection_string)
         insert_statement = "INSERT IGNORE INTO {table_name} (id) SELECT id from temp_{table_name};".format(
-                table_name=table_name)
+            table_name=table_name)
         engine.execute(insert_statement)
         engine.dispose()
 
 
-def setup_database(update_config):
+def setup_database(update_config, drop=True):
     required_tables = get_required_tables(update_config)
     print("Required tables are {tlist}".format(tlist=", ".join(required_tables)))
     connection_string = get_connection_string(update_config, "RAW_DB")
     engine = create_engine(connection_string)
     raw_database = update_config["PATENTSVIEW_DATABASES"]["RAW_DB"]
     temp_upload_database = update_config["PATENTSVIEW_DATABASES"]["TEMP_UPLOAD_DB"]
+    if drop:
+        engine.execute("""
+            DROP DATABASE if exists {temp_upload_database}
+        """.format(temp_upload_database=temp_upload_database))
     engine.execute("""
-DROP DATABASE if exists {temp_upload_database}
-    """.format(temp_upload_database=temp_upload_database))
-    engine.execute("""
-create database if not exists {temp_upload_database} default character set=utf8mb4 default collate=utf8mb4_unicode_ci
-            """.format(
-            temp_upload_database=temp_upload_database))
+            create database if not exists {temp_upload_database} default character set=utf8mb4
+             default collate=utf8mb4_unicode_ci
+        """.format(
+        temp_upload_database=temp_upload_database))
     for table in required_tables:
         print("Creating Table : {tbl}".format(tbl=table))
         con = engine.connect()
-        con.execute("drop table if exists {0}.{1}".format(temp_upload_database, table))
+        if drop:
+            con.execute("drop table if exists {0}.{1}".format(temp_upload_database, table))
         con.execute(
-                "create table if not exists {0}.{2} like {1}.{2}".format(temp_upload_database, raw_database, table))
+            "create table if not exists {0}.{2} like {1}.{2}".format(temp_upload_database, raw_database, table))
         con.close()
     engine.dispose()
 
@@ -100,8 +103,8 @@ def post_upload(**kwargs):
 
 if __name__ == '__main__':
     begin_database_setup(**{
-            "execution_date": datetime.date(2021, 2, 5)
-            })
+        "execution_date": datetime.date(2021, 2, 5)
+    })
     # upload_current_data(**{
     #         "execution_date": datetime.date(2020, 12, 1)
     #         })
